requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
django-crum==0.7.9
django-formtools==2.3
django-widget-tweaks==1.4.9
future==0.18.2
html5lib==1.1
iconic==1.0.0
Pillow==8.3.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.5
pytz==2021.3
reportlab==3.6.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
